requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
Flask-Admin==1.6.1
eventlet==0.36.1
python-socketio==5.11.4
python-engineio==4.9.1
SQLAlchemy==2.0.30
WTForms==3.1.2
python-dotenv==1.0.1
